streamlit/requirements.txt
streamlit
streamlit-option-menu
python-dotenv
pandas
plotly
